tower動かすとHP減少

+++ Assets/MarkerScripts/TowerDisp.cs
@@ -7,8 +7,8 @@ public class TowerDisp : MonoBehaviour
 {
     [Header("タワーのステータス")]
     [Tooltip("このタワーの初期ヒットポイント（HP）を設定します。")]
-    public int maxHp = 100; // HPを設定するための変数。publicなのでインスペクタに表示される
-    public int currentHp = 100;
+    public float maxHp = 100; // HPを設定するための変数。publicなのでインスペクタに表示される
+    public float currentHp = 100;
 
     // ゲームが開始された時に一度だけ呼ばれる
     void Start()
